[Draft] Annotation refactor and Cleanup
.


.
